[Rename] 함수분리
