+++ src/function/gameResultPrint.js
